+++ TowerDefense/TowerDefense/BasicTower.h
@@ -17,18 +17,18 @@ private:
 	string t_Name;
 	string t_Description;
 	int t_Cost;
-	int t_XCoord;
-	int t_YCoord;
+	float t_XCoord;
+	float t_YCoord;
 	int attackCooldown;
 	int damage;
 	
 public:
-	BasicTower(int XCoord, int YCoord);
+	BasicTower(float XCoord, float YCoord);
 
 	//Getter
 	inline int getCost() const { return t_Cost; }
-	inline int getXCoord() const { return t_XCoord; }
-	inline int getYCoord() const { return t_YCoord; }
+	inline float getXCoord() const { return t_XCoord; }
+	inline float getYCoord() const { return t_YCoord; }
 	inline int getAttackCooldown() const { return attackCooldown; }
 	inline int getDamage() const { return damage; }
 	inline string getDescription() const { return t_Description; }
