+++ TowerDefense/TowerDefense/FlameTower.cpp
@@ -2,7 +2,7 @@
 
 
 
-FlameTower::FlameTower(int XCoord, int YCoord)
+FlameTower::FlameTower(float XCoord, float YCoord)
 {
 	t_Name = "Flametower";
 	t_Description = "Ein Turm der Feuer schieﬂt";
@@ -26,8 +26,8 @@ int FlameTower::checkForEnemies(vector<DummyEnemy*>* enemyActiveVector)
 {
 	if (attackCooldown == 0) {
 		for (unsigned int i = 0; i < enemyActiveVector->size(); i++) {
-			int enemyXCoord = enemyActiveVector->at(i)->getXCoord();
-			int enemyYCoord = enemyActiveVector->at(i)->getYCoord();
+			float enemyXCoord = enemyActiveVector->at(i)->getXCoord();
+			float enemyYCoord = enemyActiveVector->at(i)->getYCoord();
 			//rechts unten
 			if ((enemyXCoord > (t_XCoord + 32)) && (enemyXCoord < (t_XCoord + 97))
 				&& (enemyYCoord > (t_YCoord + 32)) && (enemyYCoord < (t_YCoord + 97))) {
